backend/services/models/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
opencv-python-headless==4.8.1.78
Pillow==10.1.0
numpy==1.24.3
scikit-learn==1.3.2
scikit-image==0.22.0
scipy==1.11.4
redis==5.0.1
python-dotenv==1.0.0
python-multipart==0.0.6
